test: fold old unit into test_transparency
